requirements.txt
autofaiss==2.15.5
fastapi==0.88.0
faiss-cpu==1.7.3
h5py==3.8.0
img2dataset==1.41.0
jsonlines==3.1.0
loguru==0.6.0
numpy
open-clip-torch==2.16.0
pandas
prometheus-client==0.15.0
pyarrow
pydantic==1.9.1
python-dotenv==0.21.0
starlette_exporter==0.4.0
tqdm==4.64.1
transformers==4.28.0
typer[all]==0.7.0
uvicorn
